add: list all accounts

+++ week-08/day-03/bank-of-simba/src/main/java/greenfox/thymeleaf_practice/BankAccount.java
@@ -3,10 +3,10 @@ package greenfox.thymeleaf_practice;
 public class BankAccount {
 
   private String name;
-  private double balance;
+  private int balance;
   private String animalType;
 
-  public BankAccount(String name, double balance, String animalType) {
+  public BankAccount(String name, int balance, String animalType) {
     this.name = name;
     this.balance = balance;
     this.animalType = animalType;
@@ -16,7 +16,7 @@ public class BankAccount {
     return name;
   }
 
-  public double getBalance() {
+  public int getBalance() {
     return balance;
   }
